+++ src/main/java/de/geheimagentnr1/manyideas_doors/elements/blocks/big_doors/wooden/BigDoorMangrove.java
@@ -1,0 +1,51 @@
+package de.geheimagentnr1.manyideas_doors.elements.blocks.big_doors.wooden;
+
+import de.geheimagentnr1.manyideas_core.elements.block_state_properties.OpenedBy;
+import de.geheimagentnr1.manyideas_core.elements.blocks.template_blocks.doors.BigDoor;
+import de.geheimagentnr1.manyideas_doors.elements.blocks.ModBlocks;
+import net.minecraft.client.renderer.RenderType;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.level.block.SoundType;
+import net.minecraft.world.level.block.state.properties.BlockSetType;
+import net.minecraft.world.level.material.Material;
+
+
+public class BigDoorMangrove extends BigDoor {
+	
+	
+	public static final String registry_name = "big_door_mangrove";
+	
+	public BigDoorMangrove() {
+		
+		super(
+			Properties.of( Material.WOOD ).strength( 3.0F ).sound( SoundType.WOOD ),
+			BlockSetType.MANGROVE,
+			OpenedBy.BOTH,
+			true
+		);
+	}
+	
+	@Override
+	protected int getXSize() {
+		
+		return 3;
+	}
+	
+	@Override
+	protected int getYSize() {
+		
+		return 4;
+	}
+	
+	@Override
+	protected int getZSize() {
+		
+		return 3;
+	}
+	
+	@Override
+	public Item getBlockItem( Item.Properties _properties ) {
+		
+		return createBlockItem( ModBlocks.BIG_DOOR_MANGROVE, _properties, registry_name );
+	}
+}